feat(avatar): add Avatar component with props name, color, size
This commit adds a new Avatar component to the project. The Avatar component is an atom component that displays a user's avatar. It takes in three props: name, color, and size. The name prop is a string that represents the user's name. The color prop is an optional string that represents the background color of the avatar. The size prop is an optional string that represents the size of the avatar.

This commit also includes a story file for the Avatar component. The story file includes three stories: DefaultColor, CustomColor, and CustomSize. The stories demonstrate how the Avatar component can be used with different props.

The Avatar component is exported from the index file.